DetectionServer/Backend/requirements.txt
torch
torchvision
ultralytics
fastapi
pillow
uvicorn
numpy
python-multipart  # Required in container